timing_engine.py
import whisper

def get_word_times(audio):
    model = whisper.load_model("base")

    result = model.transcribe(
        audio,
        word_timestamps=True
    )

    words = []

    for seg in result["segments"]:
        for w in seg["words"]:
            words.append({
                "word": w["word"],
                "start": w["start"],
                "end": w["end"]
            })

    return words