requirements.txt
moviepy
numpy
pillow
openai-whisper
edge-tts